fix: optimize project file structure
